+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Issue [#35](https://github.com/m-marini/qucomp/issues/35): Add qubit projections
+
 ### Fixed
 
 - Issue [#33](https://github.com/m-marini/qucomp/issues/33): Wrong sqrt of negative integer